refactor: add Guest type + move src/utils/types.ts to src/types/index.ts
Also: import ts type from vue-router instead of copying it
