+++ src/store/current/points.ts
@@ -4,7 +4,7 @@
 import { Action, Module, VuexModule, getModule } from 'vuex-module-decorators'
 import store from '@/store'
 import uuid from 'uuid'
-import { Point, XYCategory, XYId } from '@/utils/types.ts'
+import { Point, XYCategory, XYId } from '@/types'
 
 import Compositions from '@/store/compositions.ts'
 const compositions = getModule(Compositions)
